html/browsers/browsing-the-web/scroll-to-fragid/fragment-and-encoding-2.html

<!doctype html>
<meta charset=windows-1252>
<title>Fragment navigation: encoding</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<div style=height:10000px></div>
<div id=&#xFFFD;></div>
<script>
test(() => {
  assert_equals(location.hash, "", "Page must be loaded with no hash");

  location.hash = "%C2";
  assert_equals(location.hash, "#%C2");
  assert_greater_than(self.scrollY, 1000, "#%C2"); // UTF-8 decoded this'll match U+FFFD

  location.hash = "top";
  assert_equals(self.scrollY, 0, "#top");

  location.hash = "%EF%BB%BF%C2";
  assert_equals(location.hash, "#%EF%BB%BF%C2");
  assert_equals(self.scrollY, 0, "#%EF%BB%BF%C2");

  location.hash = "%EF%BF%BD";
  assert_equals(location.hash, "#%EF%BF%BD");
  assert_greater_than(self.scrollY, 1000, "#%EF%BF%BD");
});
</script>
